% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndvi_match.R
\name{ndvi_match}
\alias{ndvi_match}
\title{An NDVI-coordinate matching function}
\usage{
ndvi_match(start_date, df, tiff_path_a, tiff_path_b)
}
\arguments{
\item{start_date}{a vector of dates in the form YYYY-MM-DD that correspond to Sentinel tiff files}

\item{df}{a data frame that has columns "Date," "Latitude," and "Longitude"}

\item{tiff_path_a}{string containing the first half of the path name for the tiff file before the date}

\item{tiff_path_b}{a string containing the second half of the path name for the tiff file after the date}
}
\value{
a vector of NDVIs matched to the coordinates and dates given as input
}
\description{
An NDVI-coordinate matching function
}
