% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_stats.R
\name{veg_stats}
\alias{veg_stats}
\title{A function to calculate vegetation statistics/metrics from raw vegetation sheets. Equations taken from Daniel Rubenstein.}
\usage{
veg_stats(raw_vegetation, stats_df)
}
\arguments{
\item{raw_vegetation}{a data frame containing data from vegetation transects}

\item{stats_df}{an empty header data frame to be filled out (look for raw_vegetation_header.csv)}
}
\value{
filled out stats_df
}
\description{
A function to calculate vegetation statistics/metrics from raw vegetation sheets. Equations taken from Daniel Rubenstein.
}
