% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_dist_UTM.R
\name{water_dist_UTM}
\alias{water_dist_UTM}
\title{A function to calculate distances between zebra sightings and the nearest water point}
\usage{
water_dist_UTM(df, water_loc)
}
\arguments{
\item{df}{a data frame containing metadata (presumably for zebras and cattle) and containing the columns "Date," "GPS.x," "GPS.y," and "Distance.to.water"}

\item{water_loc}{a data frame of coordinates (Arc 1960) for water locations (presumably in Mpala)}
}
\value{
a data frame identical to df but with the Distance.to.water column filled in
}
\description{
A function to calculate distances between zebra sightings and the nearest water point
}
